refactor: move algorithm functions into algorithms.h/.cpp
